automl-service/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# AutoGluon
autogluon.tabular>=1.1.0
autogluon.timeseries>=1.1.0
autogluon.multimodal>=1.1.0

# MLflow for experiment tracking
mlflow>=2.10.0

# Domino SDK
dominodatalab>=1.2.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.0
